+++ src/main/java/niv/heater/registry/HeaterRegistrar.java
@@ -9,6 +9,7 @@ import niv.heater.block.HeatPipeBlock;
 import niv.heater.block.entity.HeaterBlockEntity;
 import niv.heater.block.entity.ThermostatBlockEntity;
 
+@SuppressWarnings("null")
 @NullMarked
 public class HeaterRegistrar {
     private HeaterRegistrar() {